Port the claim arithmetic to Rust and bind it to the claim vectors
The Rust leg carries the same index and epoch as the Go one, with a zero index
meaning unset so a zeroed account is already correct. LegRecord stores the
index as two u64 halves because a u128 would force 16 byte alignment on a
zero-copy struct. Go's ClaimEpoch is now uint64 to match.

Breaking the Rust code four ways showed one hole in the shared vectors: a claim
credited again after a seizure was not covered, so a Rust build that skipped the
settle still passed. Go had a unit test for it and Rust had nothing binding it.
The Go generator now emits that scenario. The regenerated file was compared
with the committed one: the five earlier scenarios are byte identical and one is
added.

Cost: the ClaimEpoch and index add 24 bytes to every leg record.

+++ internal/basket/claim.go
@@ -27,7 +27,7 @@ type Claim struct {
 	// and nil, while Units is zero.
 	Index *big.Int
 	// Epoch is the leg's ClaimEpoch at that time.
-	Epoch uint32
+	Epoch uint64
 }
 
 // NewClaim returns an empty claim at the given scale.
